Automatic release generation
This commit adds a simple script to generate release packages for Linux
and Windows. Executing './mkrelease.sh' should now generate
ready-for-upload binary packages in Packages/. See the usage note inside
the script for details.

In the process, the launch script for Linux is replaced by one that
explicitly requests the 4.0 Mono runtime. This fixes a bug where the
version autodetection failed under certain circumstances.

fixes #2