Fix crashes in debater GUI
The debater management GUI used to crash in various situations:

1. When clicking the "Add" button with an empty list of debaters.
2. When entering non-parsable input into any of the fields with input
   restrictions, such as name, club, or age.

To fix this issue, I refactored a fairly large chunk of code. Mainly,
the EditDebaters widget, where the crashes originated, has been modified
to remove unsafe uses of reflection; and the parsing methods for basic
data containers (Name, Club, Role, DebaterPattern) have been changed to
rely on Either instead of exceptions for error reporting. Other than
that, the changes in this commit are either follow-ups to the above or
trivial improvements.

fixes #10