Remove accidental Mono.Unix dependency
Monodevelop's Stetic GUI designer attempts to use Gettext for
internationalisation unless we explicitly tell it not to. This in turn
depends on Mono.Unix, making the application non-portable to Windows.

This commit disables Gettext translation for all GUI elements and
removes the reference to Mono.Unix, though most of the changes are
just minor differences in code formatting.